Add BIP-340 compatible FROST implementation
To produce BIP-340 signatures with FROST, the H2 hash function needs to
be modified to match the BIP-340 challenge hash.

Because BIP-340 compresses curve points into X-coordinates, and defines
Y-coordinates as being even, some of the secret scalars in the protocol
execution need inverting if their corresponding Y-coordinate is odd. The
secret key d' can be inverted at key generation if the group public
point has an odd Y-coordinate; all secret key shares and corresponding
public key shares are inverted, and this inverts the group secret key.
The value k' is more difficult, and by default FROST has a 1/2 chance of
successfully producing a BIP-340 signature.