H2 function for BIP-340 hashing interface
H2 is the implementation of H2(m) function from [FROST] implemented in a way
compatible to how [BIP340] hashing functions are specified.

In the interface, we expect at least one message to be provided. Optionally, we
allow for more and they will all be concatenated. Expecting at least one message
makes sense for the hash function signature. We allow for more because it is how
this function will be used in FROST.