H3, H4, and H5 hash functions
H3, H4, and H5 functions use the tag defined in FROST. Those functions are
internal to the FROST protocol so we will use the FROST tag for a better domain
separation. This is also consistent with H1 and only H2 must use BIP340 version
given how the BIP340 verification algorithm is specified.