Call validateGroupCommitments at the beginning of Round2
There will be more functions than just encodeGroupCommitments operating
on the commitments so it makes sense to validate them at the very
beginning.

No logic in encodeGroupCommitment was altered as a part of this commit
other than the call to validateGroupCommitments was removed from this
function as it is a part of Round2 function now.