Renamed IsNotIdentity to IsPointOnCurve
For elliptic curve cryptography, we do not have an identity element but we take
the point at infinity as an identity element. The point at infinity is an extra
point O and it is not on the curve. In our case it is enough to validate if the
point we are deserializing is on the curve and we can make it clear by renaming
the function.