Implemented deriveInterpolatingValue function
deriveInterpolatingValue implements def derive_interpolating_value(L, x_i)
function from [FROST], as defined in section 4.2 Polynomials.

deriveInterpolatingValue is a simplified version of the algorithm for
computing Lagrange coefficient given we only care about the value f(0) instead
of reconstructing the entire polynomial.

L is the list of the indices of the members of the particular group.
xi is the index of the participant i.