Hardened the validation of commitments
The validateGroupCommitments logic was improved to cover the case when the
commitment from the current signer is missing. Also, it now covers the case when
one of the commitments is missing. Improved unit test coverage to check all
cases individually and then all of them together.

This allowed to simplify deriveInterpolatingValue logic and skip the validation
in this function as all the required validation happens earlier, in
validateGroupCommitments.