Extracted base commitment validation logic
The commitment validation logic that has to be performed by both signer
and coordinator has been extracted to validateGroupCommitmentsBase.

The signer performs one more validation: if the current signer's
commitment is included on the list. This validation does not have to be
performed by the coordinator (logic to be implemented) when validating
the signature shares because the coordinator will iterate over the received
commitments.