Added EncodePoint function to the Ciphersuite
EncodePoint encodes the given elliptic curve point to a byte slice in a way that
is *specific* to the given ciphersuite needs. This is especially important when
calculating a signature challenge in [FROST].

This function may yield a different result than SerializePoint function
from the Curve interface. While the SerializePoint result should be
considered an internal serialization that may be optimized for speed or
data consistency, the EncodePoint result should be considered an external
serialization, always reflecting the given ciphersuite's specification
requirements.

In the specific case of BIP-340 ciphersuite implementation, SerializePoint
serializes both X and Y coordinates, while EncodePoint serializes just the X
coordinate, as it is expected by BIP-340 for the challenge computation. The way
SerializePoint works allows to avoid computing Y coordinate manually when
exchanging data. The way EncodePoint works is dictated by BIP-340 specification.