Draft implementation of BIP-340 signature verification
This is a draft implementation. The remaining BIP-340 test vectors need to be
added, TODOs need to be addressed, and one failing test needs to be fixed.

The VerifySignature verifies the provided BIP-340 signature for the message
against the group public key. The function returns true and nil error when the
signature is valid. The function returns false and an error when the signature
is invalid. The error provides a detailed explanation on why the signature
verification failed.

VerifySignature implements Verify(pk, m, sig) function as defined in BIP-340.

One important design decision is to accept *Signature and *Point into Verify
function instead of bytes, as in the prototype. This has the implication. To
ensure consistency and that we'll not return positive verification result for
(x,y) that is not on the curve, we need to verify y coordinate even though
BIP-340 verification is not really interested in it. I think this is acceptable
given the complexity of byte operations is hidden behind a ciphersuite and
inside the protocol code, we'll operate on known domain objects.