Improved BIP340 sig verification in ciphersuite
Added missing test vectors with one test vector failing - this requires
further investigation. Explained why we accept public key's Y coordinate
in parameters as opposed to what is proposed in BIP-340. tl;dr; support
for other ciphersuites and in FROST it does not matter where we strip Y
coordinate information: after aggregate and before the verification or
inside the verification.