Fixed the expected error message in BIP340 verification test
This test was previously failing.

The signature verification algorithm in BIP-340 does not check that the 32
public key bytes in the signature match a valid point on the curve. This check
is effectively implicit in the algorithm later, because G and P are known valid
curve points, so the point R that is calculated is also a valid point. Thus, if
r is not a valid point's X-coordinate, R.X != r.