Invert secret key for an even Y coordinate of the public key
In BIP-340 there are two places in the signing protocol where we need to invert
a scalar if the elliptic curve point for that scalar has even Y coordinate. The
first place is when `d` is calculated, and the second place is where `k` is
calculated. For `d` calculation, we will have to take it into account for the
key generation protocol. In unit tests, we simply do the inversion for the
secret key, as proposed in this commit. For `k` the situation is more
complicated because it is calculated according to the FROST protocol. In unit
tests, we retry signing as proposed in this commit. For ROAST, when we'll be
building it, this is something we need to take into account as well.