Validate the number of commitments and signature shares
The number of signature shares should at least be equal to the group threshold
when aggregating them. We now validate their number to produce a clear error
message as soon as possible. Also, the number of commitments and signature
shares should be the same so that the produced signature is valid.