Port keep-network/keep-core/pkg/crypto/ephemeral
The source code of `keep-network/keep-core/pkg/crypto/ephemeral` has been
ported to this repository. The ultimate goal is to port the entire GJKR
algorithm from `keep-core` repository and parameterize it to work with any
ciphersuite. The ephemeral package is a dependency of `gjkr` package so it
had to be ported first.

To port the `ephemeral` package, I also had to port the box implementation
from `keep-network/keep-common/pkg/encryption` package. The box was only
used by `ephemeral` so I decided to simplify the code organization and place
the box implementation in the `ephemeral` package. This is the only change
compared to the original `ephemeral` implementation.