Eliminate dependency to ethereum/go-ethereum
We now use `btcsuite/btc/btcec` as a curve implementation in `ephemeral`
and `frost` packages. This allows to eliminate the dependency to
`ethereum/go-ethereum` altogether. The Marshal and Unmarshal code for points
had to be copied from `ethereum/go-ethereum` to `bip340.go`. I found
no code in `btcsuite/btc/btcec` doing what we need.