Logger interface for GJKR protocol
GJKR protocol allows the key generation to continue even if some
misbehavior inside of the group is detected. We do not want to stop
protocols returning an error yet we want to log all the misbehavior. To
not complicate the API of GJKR, a Logger interface implementation will
have to be provided when instantiating GJKR protocol members. The
current implementation is aligned with the Cosmos SDK logger interface
but the client code for the library can implements their own adapters.