GJKR message evidence log
GJKR protocol requires group members to have access to messages
exchanged between the accuser and the accused party for the stake of
complaint resolution. This is what the evidence log provides.

For now, only functions for the ephemeral public key message are
included. The code has been ported from keep-network/keep-core GJKR
implementation for the random beacon with small changes.