Detect and mark inactive members after the ephemeral pubkey generation
The phase 2 of the GJKR protocol generating symmetric keys should find
members inactive in phase 1 and mark them accordingly. I introduced this
logic in a new file, message_filter.go and moved there the deduplication
logic as well. The goal is to have all the message pre-processing for
all phases in one place. For phase 2, the pre-processing includes
deduplication, inactive member detection, and session ID handling (to be
added in the next commits). For further phases, we will also have to
filter out members marked as inactive or disqualified in previous phases
of the protocol.